// src/modules/panel/duck/panel.types.ts
export enum PanelActionTypes {
  ADD_OFFER_REQUEST = "ADD_OFFER_REQUEST",
  ADD_OFFER_SUCCESS = "ADD_OFFER_SUCCESS",
  ADD_OFFER_FAILURE = "ADD_OFFER_FAILURE",

  DELETE_OFFER_REQUEST = "DELETE_OFFER_REQUEST",
  DELETE_OFFER_SUCCESS = "DELETE_OFFER_SUCCESS",
  DELETE_OFFER_FAILURE = "DELETE_OFFER_FAILURE",

  UPDATE_USER_REQUEST = "UPDATE_USER_REQUEST",
  UPDATE_USER_SUCCESS = "UPDATE_USER_SUCCESS",
  UPDATE_USER_FAILURE = "UPDATE_USER_FAILURE"
}